+++ src/commands/constants.ts
@@ -6,3 +6,5 @@ export const AGGREGATE = 'vscode-sailpoint-identitynow.aggregate-source';
 export const AGGREGATE_DISABLE_OPTIMIZATION = 'vscode-sailpoint-identitynow.aggregate-source-disable-optimization';
 export const RESET_SOURCE = 'vscode-sailpoint-identitynow.reset-source';
 export const NEW_TRANSFORM = 'vscode-sailpoint-identitynow.new-transform';
+export const NEW_PROVISIONING_POLICY = 'vscode-sailpoint-identitynow.new-provisioning-policy';
+export const NEW_SCHEMA = 'vscode-sailpoint-identitynow.new-schema';