+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "vscode-sailpoint-identitynow" extension will be docu
 
 ## [Unreleased]
 
+### Added
+- Support for source schemas
+- Support for provision policies
+- Schema and snippets for source schemas
+- Schema and snippets for provisioning policies
+- Snippets for transforms
+
+
 [0.0.1] - 2022/02/03
 ### Added
 - Add tenant with Personal Access Token (PAT) authentication